+++ queries_test.go
@@ -1036,6 +1036,7 @@ func TestStmt_SetQueryNotification(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to open connection: %v", err)
 	}
+	defer cn.Close()
 	stmt, err := cn.Prepare("SELECT 1")
 	if err != nil {
 		t.Error("Connection failed", err)
@@ -2759,5 +2760,25 @@ func skipIfNamedPipesEnabled(t *testing.T) {
 	t.Helper()
 	if len(msdsn.ProtocolParsers) > 1 {
 		t.Skip("Skipping test due to named pipes dialer")
+	}
+}
+
+func TestAdminConnection(t *testing.T) {
+	params := testConnParams(t)
+	protocol, ok := params.Parameters["protocol"]
+	if !ok || protocol != "admin" {
+		t.Skip("Test is not running with admin protocol set")
+	}
+	conn, _ := open(t)
+	defer conn.Close()
+	row := conn.QueryRow(`SELECT  c.net_transport 
+	FROM sys.dm_exec_connections c
+	JOIN sys.tcp_endpoints e ON c.endpoint_id = e.endpoint_id
+	WHERE c.session_id = @@SPID AND e.name = 'Dedicated Admin Connection'`)
+	if err := row.Scan(&protocol); err != nil {
+		t.Fatalf("Unable to query connection protocol or not connected on DAC %s", err.Error())
+	}
+	if !strings.EqualFold(protocol, "tcp") {
+		t.Fatalf("Tcp connection not made. Protocol: %s", protocol)
 	}
 }